+++ app/src/main/AndroidManifest.xml
@@ -19,14 +19,14 @@
         android:supportsRtl="true"
         android:screenOrientation="portrait"
         android:theme="@style/MainTheme">
-        <activity android:name=".activities.StartActivity">
+        <activity android:name=".ui.activities.StartActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activities.MainActivity" />
+        <activity android:name=".ui.activities.MainActivity" />
 
     </application>
