+++ app/src/main/java/com/goryachok/forecastapp/ui/fragment/DailyForecastFragment.kt
@@ -1,4 +1,4 @@
-package com.goryachok.forecastapp.fragment
+package com.goryachok.forecastapp.ui.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,16 +6,16 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
+import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.goryachok.forecastapp.R
-import com.goryachok.forecastapp.adapters.DailyForecastAdapter
+import com.goryachok.forecastapp.ui.adapters.DailyForecastAdapter
 import com.goryachok.forecastapp.viewmodel.DailyForecastViewModel
 import kotlinx.android.synthetic.main.daily_forecast_fragment.*
 
 class DailyForecastFragment : Fragment() {
 
-    private val viewModel by lazy { ViewModelProviders.of(this).get(DailyForecastViewModel::class.java)}
+    private val viewModel by lazy { ViewModelProvider(this).get(DailyForecastViewModel::class.java) }
     private lateinit var adapter: DailyForecastAdapter
 
     override fun onCreateView(
@@ -28,12 +28,14 @@ class DailyForecastFragment : Fragment() {
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        adapter = DailyForecastAdapter(viewModel.data.value!!.list)
+        adapter =
+            DailyForecastAdapter(viewModel.data.value!!.list)
         dailyForecast_recyclerView.layoutManager = LinearLayoutManager(context)
         dailyForecast_recyclerView.adapter = adapter
-        viewModel.data.observe(this, Observer {
+        viewModel.data.observe(viewLifecycleOwner, Observer {
             adapter.update(it.list)
-            dailyForecast_cityName_textView.text=getString(R.string.location_template,it.city.name,it.city.country)
+            dailyForecast_cityName_textView.text =
+                getString(R.string.location_template, it.city.name, it.city.country)
         })
     }
 }
