+++ app/src/main/java/com/goryachok/forecastapp/viewmodel/HourlyForecastViewModel.kt
@@ -2,25 +2,9 @@ package com.goryachok.forecastapp.viewmodel
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.goryachok.forecastapp.WeatherApplication
-import com.goryachok.forecastapp.model.ForecastResponse
+import com.goryachok.forecastapp.pojo.ForecastResponse
 
 class HourlyForecastViewModel : ViewModel() {
 
     var data: MutableLiveData<ForecastResponse> = MutableLiveData()
-    private val repository by lazy { WeatherApplication.repository }
-
-    init {
-
-        data.value = repository.localForecastData.value
-
-        repository.apply {
-            lastSearchForecastData.observeForever {
-                data.postValue(it)
-            }
-            localForecastData.observeForever {
-                data.postValue(it)
-            }
-        }
-    }
 }
