+++ app/src/main/res/layout/activity_start.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimary"
-    tools:context=".activities.StartActivity">
+    tools:context=".ui.activities.StartActivity">
 
     <ImageView
         android:id="@+id/imageView"
